Patch & Bugfix

+++ src/adam/IAdam.java
@@ -9,28 +9,28 @@ public interface IAdam {
 	 * @return true - if the robot finished the movement successfully<br>
 	 * 		   false - if the robot could not finish the movement successfully.
 	 */
-	public boolean moveForward(int cm);
+	public boolean moveForward(double cm);
 	/**
 	 * Use this method to move a given length in cm backward.
 	 * @param cm
 	 * @return true - if the robot finished the movement successfully<br>
 	 * 		   false - if the robot could not finish the movement successfully.
 	 */
-	public boolean moveBackward(int cm);
+	public boolean moveBackward(double cm);
 	/**
 	 * Use this method to turn right a given angle in degree.
 	 * @param cm
 	 * @return true - if the robot finished the movement successfully<br>
 	 * 		   false - if the robot could not finish the movement successfully.
 	 */
-	public boolean turnRight(int angle);
+	public boolean turnRight(double angle);
 	/**
 	 * Use this method to turn left a given angle in degree.
 	 * @param cm
 	 * @return true - if the robot finished the movement successfully<br>
 	 * 		   false - if the robot could not finish the movement successfully.
 	 */
-	public boolean turnLeft(int angle);
+	public boolean turnLeft(double angle);
 	
 	/**
 	 * Use this method to grab and lift an item in front of the robot.
